updates to current trigger monitor
