+++ sbndqm/Decode/Trigger/detail/TriggerGateTypes.h
@@ -1,0 +1,10 @@
+namespace daq
+{
+  struct TriggerGateTypes {
+    static constexpr int BNB { 1 };
+    static constexpr int NuMI { 2 };
+    static constexpr int OffbeamBNB { 3 };
+    static constexpr int OffbeamNuMI { 4 };
+    static constexpr int Calib { 5 };
+  };
+}
