--- sbndqm/dqmAnalysis/Trigger/fcl/trigger_online_monitor_old.fcl
@@ -1,95 +1,0 @@
-#include "services_common_icarus.fcl"
-#include "simple_channel_info.fcl"
-#include "decoderdefs_icarus.fcl"
-#include "redis_connection.fcl"
-
-physics: {
-  
-  producers: {
-     daqTrigger: @local::decodeTrigger
-  }
-
-  analyzers: {
-
-   triggerAnalysis: {
-
-      module_type: TriggerStreams
-      
-      TriggerLabel: "daqTrigger"      
-      
-      TriggerMetricConfig: {
-
-        hostname: "icarus-db01.fnal.gov"
-  
-        streams: ["30s"]
-        
-        metrics: {
-	    Offbeam_BNB_RATE:{
-		units: Hz 
-		title: "Offbeam BNB triggers rate"
-		display_range:[0, 2]
-	    }
-            BNB_RATE:{
-                units: Hz
-                title: "Offbeam BNB triggers rate"
-                display_range:[0, 2]
-            }
-            Offbeam_NuMI_RATE:{
-                units: Hz
-                title: "Offbeam BNB triggers rate"
-                display_range:[0, 2]
-            }
-            NuMI_RATE:{
-                units: Hz
-                title: "Offbeam BNB triggers rate"
-                display_range:[0, 2]
-            }
-	    Unknown_RATE:{
-		units: Hz
-                title: "Offbeam BNB triggers rate"
-                display_range:[0, 2]
-            }		 
-	}   
-      }
-    }
-  }
-
-
-  my_producer_modules: [ daqTrigger ]
-  trigger_paths: [ my_producer_modules ]
-  
-  ana: [ triggerAnalysis ]
-  end_paths: [ ana ]
-
-}
-
-
-services: {
-
-  TFileService: { fileName: "waveformtest.root"}
-
-  DetectorClocksService: @local::icarus_detectorclocks
-
-  RedisConnection: @local::redis_connection
-
-  InitMetricManager: {
-    connections: {
-      fast: @local::redis_metric_fast
-      slow: @local::redis_metric_slow
-      archive: @local::redis_metric_archive
-    }
-  }
-
-} 
-
-
-services.RedisConnection.host: "icarus-db01.fnal.gov"
-
-source:
-{
-  module_type: RootInput
-  fileNames: ["../data/data_dl1_fstrmNUMI_run6983_8_20211028T113506.root"]
-}
-
-
-process_name: ICARUSTRIGGERANALYSIS
